requirements.txt
alembic==1.8.0
bcrypt==3.2.2
cachelib==0.9.0
cffi==1.15.0
click==8.0.3
colorama==0.4.5
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.1
Flask-Migrate==3.1.0
Flask-MySQLdb==1.0.1
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
protobuf==4.21.1
pycparser==2.21
PyJWT==2.3.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
six==1.16.0
SQLAlchemy==1.4.39
Werkzeug==2.1.2
WTForms==3.0.1
zipp==3.8.0